move messageBox functions to a new submodule
